+++ src/main/java/org/klodnicki/DTO/Employee/EmployeeDTORequest.java
@@ -1,4 +1,4 @@
-package org.klodnicki.DTO;
+package org.klodnicki.DTO.Employee;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
